document-intelligence-refinery/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "Document Intelligence Refinery MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic==2.11.1",
  "pdfplumber==0.11.5",
  "pymupdf==1.25.3",
  "pyyaml==6.0.2",
  "fasttext-langdetect==1.0.5",
  "docling==2.30.0",
  "openai==1.68.2",
  "fastapi==0.115.8",
  "python-multipart>=0.0.9",
  "uvicorn==0.34.0",
  "python-dotenv>=1.0.0",
  "langchain==0.3.19",
  "langgraph==0.2.74",
  "langsmith==0.3.8",
  "chromadb==0.6.3",
  "sentence-transformers>=2.2.0",
  "faiss-cpu==1.10.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-asyncio==0.25.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "document_refinery*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
